telegram-bot/requirements.txt
aiogram==3.0.0b7
confluent-kafka==2.2.0
python-dotenv==1.0.0